Miscellaneous

Database

DearDiary.initialize_database — Function

initialize_database(; file_name::String="deardiary.db")

Initializes the database by creating the necessary tables.

DearDiary.get_database — Function

get_database()::Union{SQLite.DB,Nothing}

Returns a SQLite database connection. If the database has not been initialized, it returns nothing.

Returns

A SQLite.DB object, or nothing if the database is not initialized.

DearDiary.close_database — Function

close_database()

Closes the database connection if it is open.

Enumerations

DearDiary.Status — Type

@enum Status IN_PROGRESS = 1 STOPPED = 2 FINISHED = 3

An enumeration representing the status of an experiment.

DearDiary.IN_PROGRESS — Constant

IN_PROGRESS::Status

An enumeration value representing an experiment that is currently in progress.

DearDiary.STOPPED — Constant

STOPPED::Status

An enumeration value representing an experiment that has been stopped.

DearDiary.FINISHED — Constant

FINISHED::Status

An enumeration value representing an experiment that has finished.

Marker types

DearDiary.UpsertResult — Type

UpsertResult

A marker abstract type for the result of an upsert operation.

DearDiary.Created — Type

Created

A marker type indicating that a record was successfully created.

DearDiary.Updated — Type

Updated <: UpsertResult

A marker type indicating that a record was successfully updated.

DearDiary.Duplicate — Type

Duplicate <: UpsertResult

A marker type indicating that a record already exists.

DearDiary.Unprocessable — Type

Unprocessable <: UpsertResult

A marker type indicating that a record violates a constraint and cannot be processed.

DearDiary.Error — Type

Error <: UpsertResult

A marker type indicating that an error occurred while creating or updating a record.