Types

DearDiary.User — Type

User <: ResultType

A struct that represents a user.

Fields

id::Int64: The ID of the user.

first_name::String: The first name of the user.

last_name::String: The last name of the user.

username::String: The username of the user.

password::String: The password of the user. This is a hashed version of the password, not the plain text password.

created_date::DateTime: The date and time the user was created.

is_admin::Bool: Whether the user is an administrator.

DearDiary.UserPermission — Type

UserPermission <: ResultType

A struct representing a user's permissions for a specific project.

Fields

id::Int64: The unique identifier for the user permission record.

user_id::Int64: The ID of the user.

project_id::Int64: The ID of the project.

create_permission::Bool: Permission to create resources.

read_permission::Bool: Permission to read resources.

update_permission::Bool: Permission to update resources.

delete_permission::Bool: Permission to delete resources.

DearDiary.Project — Type

Project <: ResultType

A struct representing a project with its details.

Fields

id::Int64: The ID of the project.

name::String: The name of the project.

description::String: A brief description of the project.

created_date::DateTime: The date and time the project was created.

DearDiary.Experiment — Type

Experiment <: ResultType

A struct representing an experiment within a project.

Fields

id::Int64: The unique identifier of the experiment.

project_id::Int64: The identifier of the project to which the experiment belongs.

status_id::Int64: The status of the experiment.

name::String: The name of the experiment.

description::String: A description of the experiment.

created_date::DateTime: The date and time when the experiment was created.

end_date::Optional{DateTime}: The date and time when the experiment ended, or nothing if it is still ongoing.

DearDiary.Iteration — Type

Iteration <: ResultType

A struct representing an iteration within an experiment.

Fields

id::Int64: The unique identifier of the iteration.

experiment_id::Int64: The identifier of the experiment to which the iteration belongs.

notes::String: Notes associated with the iteration.

created_date::DateTime: The date and time when the iteration was created.

end_date::Optional{DateTime}: The date and time when the iteration ended, or nothing if it is still ongoing.

DearDiary.Parameter — Type

Parameter <: ResultType

A struct representing a parameter with its details.

Fields

id::Int64: The ID of the parameter.

iteration_id::Int64: The ID of the iteration this parameter belongs to.

key::String: The key/name of the parameter.

value::String: The value of the parameter.

DearDiary.Metric — Type

Metric <: ResultType

A struct representing a metric with its details.

Fields

id::Int64: The ID of the metric.

iteration_id::Int64: The ID of the iteration this metric belongs to.

key::String: The key/name of the metric.

value::Float64: The value of the metric.

DearDiary.Resource — Type

Resource <: ResultType

A struct representing a resource associated with an experiment.

Fields

id::Int64: The ID of the resource.

experiment_id::Int64: The ID of the experiment this resource belongs to.

name::String: The name of the resource.

description::String: A description of the resource.

data::Optional{Array{UInt8,1}}: The binary data of the resource.

created_date::DateTime: The date and time when the resource was created.

updated_date::Optional{DateTime}: The date and time when the resource was last updated.