Tracking Results

Tracking Results

Tracking.TrackingResults — Type.

TrackingResults that hold the correlation results.

source

Tracking.get_early — Method.

get_early(results)

Get the early of the tracking result.

source

Tracking.get_prompt — Method.

get_prompt(results)

Get the prompt of the tracking result.

source

Tracking.get_late — Method.

get_late(results)

Get the late of the tracking result.

source

Tracking.get_correlator — Function.

get_correlator(results)

Get the correlator of the tracking result.

source

Tracking.get_carrier_doppler — Function.

get_carrier_doppler(results)

Get the carrier doppler of the tracking result.

source

Tracking.get_carrier_phase — Function.

get_carrier_phase(results)

Get the carrier phase of the tracking result.

source

Tracking.get_code_doppler — Function.

get_code_doppler(results)

Get the code doppler of the tracking result.

source

Tracking.get_code_phase — Function.

get_code_phase(results)

Get the code phase of the tracking result.

source

Tracking.get_secondary_code_or_bit_found — Function.

get_secondary_code_or_bit_found(results)

Check if the secondary code or bit has been found.

source

Tracking.get_state — Function.

get_state(results)

Get state of the tracking result.

source

Tracking.get_cn0 — Function.

get_cn0(results)

Get the carrier to noise density ratio of the tracking result.

source