Track

Tracking.track — Function

track(signal, state, prn, sample_frequency; post_corr_filter, intermediate_frequency, max_integration_time, min_integration_time, early_late_sample_shift, carrier_loop_filter_bandwidth, code_loop_filter_bandwidth, velocity_aiding)

Track the signal signal based on the current tracking state, the sample frequency sample_frequency and PRN prn. Optional configurations are:

Post correlation filter post_corr_filter defaults to get_default_post_corr_filter(...)

Intermediate frequency intermediate_frequency defaults to 0Hz

Maximal integration time max_integration_time defaults to 1ms. The actual integration time can be less in order to find the secondary code or the bit shift. Once this is found the integration time is identical to the maximal integration time.

Minimal integration time min_integration_time defaults to 0.75ms. It's the minimal integration time, that leads to a valid correlation result. It is only used for the first integration period.

Sample shift between early and late early_late_sample_shift defaults to get_early_late_sample_shift(...)

Bandwidth of the carrier loop carrier_loop_filter_bandwidth defaults to 18Hz

Bandwidth of the code loop code_loop_filter_bandwidth defaults to 1Hz

Velocity aiding velocity_aiding defaults to 0Hz

Tracking.get_early_late_sample_shift — Function

get_early_late_sample_shift(?, correlator, sample_frequency, preferred_code_shift)

Calculate the shift between the early and late in samples.